Little Alien
- removed some files
- bug fix: breath reduction while raining





git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@118 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ GameSettings.h
@@ -215,7 +215,7 @@ typedef struct
 	UINT16 gusRainMaxLength;
 	UINT32 guiMaxRainDrops;
 	UINT32 ubWeaponReliabilityReductionPerRainIntensity;
-	UINT32 ubBreathGainReductionPerRainIntensity;
+	double ubBreathGainReductionPerRainIntensity;
 
 	// Thunder settings
 	UINT32 guiMinLightningInterval;
